requirements.txt
# # Dependencies
# onnx
# onnxruntime
# ultralytics
# fastapi[all]
# gunicorn
# aiofiles
# pytest

# 为了减少构建镜像的时间，这里把版本锁定了，就不会去下载最新的版本了。
# torch 在 Dockerfile 中提前安装了，主要是要安装 CPU 版本，不然在这里执行会安装 GPU 版本，最终还会报错。
# torch==2.0.0
# torchvision==0.15.1
aiofiles==23.1.0
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.1
fastapi==0.95.0
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanfriendly==10.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pluggy==1.0.0
protobuf==3.20.3
psutil==5.9.4
pydantic==1.10.7
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.19.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
thop==0.1.1.post2209072238
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.68
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1